+++ domain/usecase/src/main/java/co/com/pragma/usecase/user/user/FindUserUseCaseInterface.java
@@ -4,9 +4,7 @@ import co.com.pragma.model.user.User;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
-public interface UserUseCaseInterface {
-    public Mono<User> saveUser(User user);
-    public Mono<User> updateUser(User user);
+public interface FindUserUseCaseInterface {
     public Flux<User> getAllUsers();
     public Mono<User> getUserByIdentityNumber(String identityNumber);
 }
