+++ domain/usecase/src/main/java/co/com/pragma/usecase/user/user/SaveUserUseCaseInterface.java
@@ -1,0 +1,9 @@
+package co.com.pragma.usecase.user.user;
+
+import co.com.pragma.model.user.User;
+import reactor.core.publisher.Flux;
+import reactor.core.publisher.Mono;
+
+public interface SaveUserUseCaseInterface {
+    public Mono<User> execute(User user);
+}
