+++ domain/usecase/src/main/java/co/com/pragma/usecase/user/user/UpdateUserUseCaseInterface.java
@@ -1,0 +1,8 @@
+package co.com.pragma.usecase.user.user;
+
+import co.com.pragma.model.user.User;
+import reactor.core.publisher.Mono;
+
+public interface UpdateUserUseCaseInterface {
+    public Mono<User> execute(User user);
+}
