+++ domain/usecase/src/test/java/co/com/pragma/usecase/user/SaveUserUseCaseTest.java
@@ -4,7 +4,7 @@ import co.com.pragma.model.role.Role;
 import co.com.pragma.model.role.gateways.RoleRepository;
 import co.com.pragma.model.user.User;
 import co.com.pragma.model.user.gateways.UserRepository;
-import co.com.pragma.usecase.user.user.UserUseCase;
+import co.com.pragma.usecase.user.user.SaveUserUseCase;
 import co.com.pragma.usecase.user.user.validations.error.UserValidationException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -16,16 +16,16 @@ import reactor.test.StepVerifier;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.*;
 
-class UserUseCaseTest {
+class SaveUserUseCaseTest {
     private UserRepository userRepository;
     private RoleRepository roleRepository;
-    private UserUseCase userUseCase;
+    private SaveUserUseCase saveUserUseCase;
 
     @BeforeEach
     void setUp() {
         userRepository = mock(UserRepository.class);
         roleRepository = mock(RoleRepository.class);
-        userUseCase = new UserUseCase(userRepository, roleRepository);
+        saveUserUseCase = new SaveUserUseCase(userRepository, roleRepository);
     }
 
     @Test
@@ -53,7 +53,7 @@ class UserUseCaseTest {
                     return Mono.just(u);
                 });
 
-        StepVerifier.create(userUseCase.saveUser(user))
+        StepVerifier.create(saveUserUseCase.execute(user))
                 .expectNextMatches(u -> u.getName().equals("Juan") &&
                         u.getRoleId() != null &&
                         u.getRoleId().equals(1L))
@@ -64,97 +64,6 @@ class UserUseCaseTest {
         User savedUser = captor.getValue();
         assertEquals("Juan", savedUser.getName());
         assertEquals(1L, savedUser.getRoleId());
-    }
-
-    @Test
-    void shouldEditAnUser() {
-        User user = User.builder()
-                .id(1L)
-                .name("Juan")
-                .lastname("Ceballos")
-                .email("juan@mail.com")
-                .identityNumber("123456")
-                .baseSalary(3000000L)
-                .build();
-
-        Role role = new Role(1L, "ADMIN", "Administrador");
-
-        when(userRepository.findByEmail("juan@mail.com"))
-                .thenReturn(Mono.empty());
-        when(roleRepository.findById(any()))
-                .thenReturn(Mono.just(role));
-        when(userRepository.save(user))
-                .thenAnswer(invocation -> {
-                    User u = invocation.getArgument(0);
-                    u.setRoleId(role.getId());
-                    return Mono.just(u);
-                });
-
-        user.setName("Juanjo");
-
-        StepVerifier.create(userUseCase.updateUser(user))
-                .expectNextMatches(u -> u.getName().equals("Juanjo") &&
-                        u.getRoleId() != null &&
-                        u.getRoleId().equals(1L))
-                .verifyComplete();
-
-        ArgumentCaptor<User> captor = ArgumentCaptor.forClass(User.class);
-        verify(userRepository).save(captor.capture());
-        User savedUser = captor.getValue();
-        assertEquals("Juanjo", savedUser.getName());
-        assertEquals(1L, savedUser.getRoleId());
-    }
-
-    @Test
-    void shouldFindAllUsers() {
-        User user1 = User.builder()
-                .id(1L)
-                .name("Juan")
-                .lastname("Ceballos")
-                .email("juan@mail.com")
-                .identityNumber("123456")
-                .baseSalary(3000000L)
-                .build();
-
-        User user2 = User.builder()
-                .id(1L)
-                .name("Leidy")
-                .lastname("Pelaez")
-                .email("leidy@mail.com")
-                .identityNumber("654321")
-                .baseSalary(3000000L)
-                .build();
-
-        when(userRepository.findAll()).thenReturn(Flux.just(user1, user2));
-
-        StepVerifier.create(userUseCase.getAllUsers())
-                .expectNext(user1)
-                .expectNext(user2)
-                .verifyComplete();
-
-        verify(userRepository, times(1)).findAll();
-
-    }
-
-    @Test
-    void shouldFindUserByIdentityNumber() {
-        String identityNumber = "123";
-        User user = User.builder()
-                .id(1L)
-                .name("Juan")
-                .lastname("Ceballos")
-                .email("juan@mail.com")
-                .identityNumber(identityNumber)
-                .baseSalary(3000000L)
-                .build();
-
-        when(userRepository.findByIdentityNumber(identityNumber)).thenReturn(Mono.just(user));
-
-        StepVerifier.create(userUseCase.getUserByIdentityNumber(identityNumber))
-                .expectNext(user)
-                .verifyComplete();
-
-        verify(userRepository, times(1)).findByIdentityNumber(identityNumber);
     }
 
     @Test
@@ -176,7 +85,7 @@ class UserUseCaseTest {
         when(roleRepository.findById(any()))
                 .thenReturn(Mono.just(role));
 
-        StepVerifier.create(userUseCase.saveUser(user))
+        StepVerifier.create(saveUserUseCase.execute(user))
                 .expectErrorMatches(throwable -> throwable instanceof UserValidationException &&
                         throwable.getMessage().equals("El campo email debe tener un formato válido."))
                 .verify();
@@ -201,7 +110,7 @@ class UserUseCaseTest {
         when(roleRepository.findById(any()))
                 .thenReturn(Mono.empty());
 
-        StepVerifier.create(userUseCase.saveUser(user))
+        StepVerifier.create(saveUserUseCase.execute(user))
                 .expectErrorMatches(throwable -> throwable instanceof UserValidationException &&
                         throwable.getMessage().equals("El Id del rol suministrado no existe."))
                 .verify();
